# multas-backend/docs/API_OPENAPI.yaml
openapi: 3.1.0
info:
  title: PMT Multas API - TANDAS 0–8, 10 y 11
  version: 1.2.0
  description: Contrato de autenticación, operación municipal, finanzas, solvencias, analítica, notificaciones y migración histórica controlada.
servers:
  - url: /api/v1
...
paths:
...
  /admin/historical-migrations/uploads:
    post:
      tags: [Historical migrations]
      security: [{ sessionCookie: [] }]
      summary: Almacena un CSV en staging privado; no escribe tablas operativas
      parameters:
        - { in: header, name: X-File-Name, required: true, schema: { type: string, pattern: '^[^/\\]+\.csv$' } }
        - { in: header, name: X-Source-System, required: true, schema: { type: string, minLength: 2, maxLength: 100 } }
        - { in: header, name: X-Migration-Entity, required: true, schema: { $ref: '#/components/schemas/HistoricalEntity' } }
      requestBody:
        required: true
        content:
          text/csv: { schema: { type: string, format: binary } }
      responses:
        '201': { description: Nuevo lote UPLOADED }
        '200': { description: Reintento idempotente del mismo archivo }
        '403': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
        '415': { $ref: '#/components/responses/Error' }
  /admin/historical-migrations/batches:
    get:
      tags: [Historical migrations]
      security: [{ sessionCookie: [] }]
      summary: Lista lotes históricos paginados
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: entity, schema: { $ref: '#/components/schemas/HistoricalEntity' } }
      responses: { '200': { description: Lotes y total }, '403': { $ref: '#/components/responses/Error' } }
  /admin/historical-migrations/batches/{id}:
    get:
      tags: [Historical migrations]
      security: [{ sessionCookie: [] }]
      summary: Consulta resumen, archivo y bitácora del lote
      parameters: [{ $ref: '#/components/parameters/MigrationBatchId' }]
      responses: { '200': { description: Detalle de lote }, '404': { $ref: '#/components/responses/Error' } }
  /admin/historical-migrations/batches/{id}/validate:
    post:
      tags: [Historical migrations]
      security: [{ sessionCookie: [] }]
      summary: Valida staging, duplicados, relaciones y catálogos sin escribir operación
      parameters: [{ $ref: '#/components/parameters/MigrationBatchId' }]
      responses: { '200': { description: Resultado VALIDATED o REQUIRES_REVIEW }, '409': { $ref: '#/components/responses/Error' } }
...
  /admin/historical-migrations/batches/{id}/approve:
    post:
      tags: [Historical migrations]
      security: [{ sessionCookie: [] }]
      summary: Lleva un lote VALIDATED sin hallazgos a READY
      parameters: [{ $ref: '#/components/parameters/MigrationBatchId' }]
      responses: { '200': { description: Lote READY }, '409': { $ref: '#/components/responses/Error' } }
...
  /admin/historical-migrations/batches/{id}/report:
    get:
      tags: [Historical migrations]
      security: [{ sessionCookie: [] }]
      summary: Descarga conciliación CSV del lote
      parameters: [{ $ref: '#/components/parameters/MigrationBatchId' }]
      responses: { '200': { description: Reporte CSV, content: { text/csv: { schema: { type: string } } } } }
  /admin/historical-migrations/templates/{entity}:
    get:
      tags: [Historical migrations]
      security: [{ sessionCookie: [] }]
      summary: Descarga una plantilla CSV con ejemplo sintético
      parameters:
        - { in: path, name: entity, required: true, schema: { $ref: '#/components/schemas/HistoricalEntity' } }
      responses: { '200': { description: Plantilla CSV, content: { text/csv: { schema: { type: string } } } } }
components:
  parameters:
    MigrationBatchId:
      in: path
      name: id
      required: true
      schema: { oneOf: [{ type: integer, minimum: 1 }, { type: string, format: uuid }] }
  schemas:
    HistoricalEntity:
      type: string
      enum: [citizens, vehicles, agents, infractions, infraction_items, payments, payment_allocations, receipts, adjustments, solvencies]
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pmt_session
  responses:
    Error:
      description: Error estable con requestId
      content:
        application/json:
          schema:
            type: object
            required: [error, meta]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
              meta:
                type: object
                required: [requestId]
                properties:
                  requestId: { type: string, format: uuid }