# multas-backend/src/modules/historical-migrations/http/historical-migration.routes.ts
/* eslint-disable @typescript-eslint/no-non-null-assertion */
import { readFile } from "node:fs/promises";
import { Router } from "express";
import { z } from "zod";
import type { AppContainer } from "../../../bootstrap/container.js";
import { authenticate } from "../../../shared/http/authenticate.js";
import { authorize } from "../../../shared/http/authorize.js";
import { HttpError } from "../../../shared/http/HttpError.js";
import { getRequestId } from "../../../shared/http/request-context.js";
import { recordOperation } from "../../../shared/http/operations.js";
import { historicalEntities } from "../application/HistoricalCsv.js";

const entity=z.enum(historicalEntities);
const id=z.union([z.coerce.number().int().positive(),z.uuid()]);
const paging=z.object({page:z.coerce.number().int().min(1).default(1),pageSize:z.coerce.number().int().min(1).max(100).default(25)});
const listing=paging.extend({status:z.string().trim().max(40).optional(),entity:entity.optional()});
const confirmation=z.object({confirmation:z.string().trim().min(1).max(100)});
const mapping=z.object({mappingType:z.enum(["TABLE","COLUMN","STATUS","ARTICLE","AGENT","PAYMENT_METHOD"]),entityType:entity.optional(),sourceValue:z.string().trim().min(1).max(191),targetEntityType:z.string().trim().min(1).max(80).optional(),targetEntityId:z.coerce.number().int().positive().optional(),targetValue:z.string().trim().min(1).max(500).optional(),conflictId:z.coerce.number().int().positive().optional()});

export function createHistoricalMigrationRouter(container:AppContainer):Router{
  const router=Router();
  router.use(authenticate(container.authRepository,container.env));
  router.use(async(request,_response,next)=>{try{if(!request.auth?.user.roles.includes("ADMIN")){if(request.auth)await container.auditRepository.record({actorUserId:request.auth.user.id,actorSessionId:request.auth.id,action:"ACCESS_DENIED",module:"historical_migrations",entityType:"role",entityId:"ADMIN",outcome:"DENIED",reason:"ADMIN_ROLE_REQUIRED",requestId:getRequestId(),ipAddress:request.ip??null,userAgent:request.get("user-agent")?.slice(0,500)??null});throw new HttpError({code:"AUTH_FORBIDDEN",message:"Esta operación requiere un administrador autorizado.",statusCode:403});}next();}catch(error){next(error);}});
  router.post("/uploads",authorize("historical_migrations.upload",container.auditRepository),async(request,response,next)=>{try{if(!Buffer.isBuffer(request.body))throw new HttpError({code:"MIGRATION_CSV_BODY_REQUIRED",message:"Envíe el CSV como cuerpo binario con Content-Type text/csv.",statusCode:415});const migrationEntity=entity.parse(request.get("x-migration-entity"));const sourceSystem=z.string().trim().min(2).max(100).parse(request.get("x-source-system"));const originalName=z.string().trim().min(1).max(255).parse(request.get("x-file-name"));const result=await container.historicalMigrations.upload({content:request.body,originalName,mimeType:request.get("content-type")??"",sourceSystem,entity:migrationEntity,actorUserId:request.auth!.user.id});await recordOperation(container,request,{action:"HISTORICAL_MIGRATION_UPLOADED",module:"historical_migrations",entityType:"migration_batch",entityId:String(result.batch.id),newValues:{entity:migrationEntity,sourceSystem}});response.status(result.idempotentReplay?200:201).json({data:result,meta:{requestId:getRequestId()}});}catch(error){next(error);}});
  router.get("/templates/:entity",authorize("historical_migrations.read",container.auditRepository),async(request,response,next)=>{try{const value=entity.parse(request.params["entity"]);response.type("text/csv").attachment(`${value}.csv`).send(await readFile(container.historicalMigrations.templatePath(value)));}catch(error){next(error);}});
  router.get("/batches",authorize("historical_migrations.read",container.auditRepository),async(request,response,next)=>{try{const query=listing.parse(request.query);const input={page:query.page,pageSize:query.pageSize,...(query.status===undefined?{}:{status:query.status}),...(query.entity===undefined?{}:{entity:query.entity})};response.json({data:await container.historicalMigrations.list(input),meta:{requestId:getRequestId()}});}catch(error){next(error);}});
  router.get("/batches/:id",authorize("historical_migrations.read",container.auditRepository),async(request,response,next)=>{try{response.json({data:await container.historicalMigrations.detail(id.parse(request.params["id"])),meta:{requestId:getRequestId()}});}catch(error){next(error);}});
  router.post("/batches/:id/validate",authorize("historical_migrations.validate",container.auditRepository),async(request,response,next)=>{try{const batchId=id.parse(request.params["id"]);const result=await container.historicalMigrations.validate(batchId,request.auth!.user.id,getRequestId());await recordOperation(container,request,{action:"HISTORICAL_MIGRATION_VALIDATED",module:"historical_migrations",entityType:"migration_batch",entityId:String(batchId)});response.json({data:result,meta:{requestId:getRequestId()}});}catch(error){next(error);}});
  for(const [path,method] of [["preview","preview"],["errors","errors"],["conflicts","conflicts"]] as const)router.get(`/batches/:id/${path}`,authorize("historical_migrations.read",container.auditRepository),async(request,response,next)=>{try{const query=paging.parse(request.query);response.json({data:await container.historicalMigrations[method](id.parse(request.params["id"]),query.page,query.pageSize),meta:{requestId:getRequestId()}});}catch(error){next(error);}});
  router.patch("/batches/:id/mappings",authorize("historical_migrations.map",container.auditRepository),async(request,response,next)=>{try{const batchId=id.parse(request.params["id"]);const value=mapping.parse(request.body);const input={mappingType:value.mappingType,sourceValue:value.sourceValue,...(value.entityType===undefined?{}:{entityType:value.entityType}),...(value.targetEntityType===undefined?{}:{targetEntityType:value.targetEntityType}),...(value.targetEntityId===undefined?{}:{targetEntityId:value.targetEntityId}),...(value.targetValue===undefined?{}:{targetValue:value.targetValue}),...(value.conflictId===undefined?{}:{conflictId:value.conflictId})};const result=await container.historicalMigrations.saveMapping(batchId,input,request.auth!.user.id,getRequestId());await recordOperation(container,request,{action:"HISTORICAL_MIGRATION_MAPPING_SAVED",module:"historical_migrations",entityType:"migration_batch",entityId:String(batchId)});response.json({data:result,meta:{requestId:getRequestId()}});}catch(error){next(error);}});
  router.post("/batches/:id/approve",authorize("historical_migrations.approve",container.auditRepository),async(request,response,next)=>{try{const batchId=id.parse(request.params["id"]);const result=await container.historicalMigrations.approve(batchId,request.auth!.user.id,getRequestId());await recordOperation(container,request,{action:"HISTORICAL_MIGRATION_APPROVED",module:"historical_migrations",entityType:"migration_batch",entityId:String(batchId)});response.json({data:result,meta:{requestId:getRequestId()}});}catch(error){next(error);}});
  for(const [path,permission,action,method] of [["import","historical_migrations.import","HISTORICAL_MIGRATION_IMPORTED","commit"],["revert","historical_migrations.revert","HISTORICAL_MIGRATION_REVERTED","revert"]] as const)router.post(`/batches/:id/${path}`,authorize(permission,container.auditRepository),async(request,response,next)=>{try{const batchId=id.parse(request.params["id"]);const input=confirmation.parse(request.body);const result=await container.historicalMigrations[method](batchId,request.auth!.user.id,input.confirmation,getRequestId());await recordOperation(container,request,{action,module:"historical_migrations",entityType:"migration_batch",entityId:String(batchId)});response.json({data:result,meta:{requestId:getRequestId()}});}catch(error){next(error);}});
  router.get("/batches/:id/report",authorize("historical_migrations.read",container.auditRepository),async(request,response,next)=>{try{const batchId=id.parse(request.params["id"]);response.type("text/csv").attachment(`migration-${String(batchId)}-report.csv`).send(await container.historicalMigrations.reportCsv(batchId));}catch(error){next(error);}});
  return router;
}